front: make enhancedEditoastApi the default
We use @rtk-query/codegen-openapi to automatically generate types
and methods from our OpenAPI schema. This works pretty well in
general, however for some specific endpoints we might need to
adjust the generated code a bit, for instance to slightly transform
the response or override the invalidatesTags field.

Right now the generated code lives in osrdEditoastApi.ts, and the
adjustments live in enhancedEditoastApi.ts. Other files import
osrdEditoastApi.ts in general, except when they need the enhanced
API, in which case they import enhancedEditoastApi.ts instead. This
is a bit cumbersome, because the imports need to be changed all
over the place when enhancing an endpoint, and it's easy to import
the wrong version in new code.

Move the generated code to a new generatedEditoastApi.ts file, which
is never directly imported by users. Move the enhanced API to
osrdEditoastApi.ts, which is the file imported by almost all users.
That way, everybody always gets the enhanced API.

The new osrdEditoastApi.ts file needs to re-export all of the
generated types alongside its own osrdEditoastApi. This creates a
conflict: there are two osrdEditoastApi symbols (one from the
re-exported generatedEditoastApi.ts, another for the enhanced API).
According to the ECMAScript standard [1], the named export should
get priority over the wildcard re-export. However we found that
in practice on some setups the wrong osrdEditoastApi was being
exported. Avoid this problem altogether by changing the name of
the generated API from osrdEditoastApi to generatedEditoastApi.

Add a small stub in enhancedEditoastApi.ts to just forward the
enhanced API from osrdEditoastApi.ts to not break existing users.

[1]: https://262.ecma-international.org/6.0/#sec-getexportednames
