+++ front/scripts/generate-types.sh
@@ -4,7 +4,7 @@ if [ $? -ne 0 ]; then
     echo "npx @rtk-query/codegen-openapi src/config/openapi-editoast-config.ts command failed. Exit the script"
     exit 1
 fi
-eslint --fix src/common/api/osrdEditoastApi.ts --no-ignore
+eslint --fix src/common/api/generatedEditoastApi.ts --no-ignore
 npx @rtk-query/codegen-openapi src/config/openapi-gateway-config.ts
 if [ $? -ne 0 ]; then
     echo "npx @rtk-query/codegen-openapi src/config/openapi-gateway-config.ts command failed. Exit the script"
